:muscle: Move service related modules into `service`
